move FileLock to own submodule
